#2047 extract arguments printing logic to separate class ArgumentsPrinter
