+++ src/main/java/com/codeborne/selenide/logevents/ArgumentsPrinter.java
@@ -1,0 +1,85 @@
+package com.codeborne.selenide.logevents;
+
+import com.codeborne.selenide.impl.DurationFormat;
+
+import javax.annotation.CheckReturnValue;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.time.Duration;
+import java.util.Arrays;
+import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.joining;
+
+public class ArgumentsPrinter {
+  private static final DurationFormat df = new DurationFormat();
+
+  @CheckReturnValue
+  @Nonnull
+  @SuppressWarnings("ChainOfInstanceofChecks")
+  static String readableArguments(@Nullable Object... args) {
+    if (args == null || args.length == 0) {
+      return "";
+    }
+
+    if (args[0] instanceof Object[]) {
+      return arrayToString((Object[]) args[0]);
+    }
+
+    if (args[0] instanceof int[]) {
+      return arrayToString((int[]) args[0]);
+    }
+
+    return arrayToString(args);
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  private static String arrayToString(Object[] args) {
+    Object[] argsWithoutEmptyVarargs = argsWithoutEmptyVarargs(args);
+    return switch (argsWithoutEmptyVarargs.length) {
+      case 0 -> "";
+      case 1 -> argToString(argsWithoutEmptyVarargs[0]);
+      default -> '[' + Stream.of(argsWithoutEmptyVarargs).map(ArgumentsPrinter::argToString).collect(joining(", ")) + ']';
+    };
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  private static Object[] argsWithoutEmptyVarargs(Object[] args) {
+    if (args.length == 0) return args;
+    Object last = args[args.length - 1];
+    if (last == null || !last.getClass().isArray()) return args;
+
+    Object[] vararg = (Object[]) last;
+    if (vararg.length == 0) {
+      return Arrays.copyOf(args, args.length - 1);
+    }
+    if (args.length < 2 || args[args.length - 2].getClass() == vararg.getClass().getComponentType()) {
+      Object[] mergedArgs = new Object[args.length - 1 + vararg.length];
+      System.arraycopy(args, 0, mergedArgs, 0, args.length - 1);
+      System.arraycopy(vararg, 0, mergedArgs, args.length - 1, vararg.length);
+      return mergedArgs;
+    }
+    return args;
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  @SuppressWarnings("ChainOfInstanceofChecks")
+  private static String argToString(Object arg) {
+    if (arg instanceof Duration) {
+      return df.format((Duration) arg);
+    }
+    if (arg instanceof Object[] array) {
+      return Arrays.toString(array);
+    }
+    return String.valueOf(arg);
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  private static String arrayToString(int[] args) {
+    return args.length == 1 ? String.valueOf(args[0]) : Arrays.toString(args);
+  }
+}
